@@ -1,7 +1,7 @@
 # Official WooSignal App Template for WooCommerce
 
 # Label StoreMax
-# Version: 6.4.0
+# Version: 6.4.1
 # Author: Anthony Gordon
 # Homepage: https://woosignal.com
 # Documentation: https://woosignal.com/docs/app/label-storemax
@@ -22,16 +22,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.3
   analyzer: ^4.2.0
   intl: ^0.17.0
-  nylo_framework: ^4.0.0
-  woosignal: ^3.2.1
-  flutter_stripe: ^7.0.0
-  wp_json_api: ^3.3.1
+  nylo_framework: ^4.1.3
+  woosignal: ^3.2.3
+  flutter_stripe: ^8.0.0+1
+  wp_json_api: ^3.3.2
   cached_network_image: ^3.2.3
   package_info: ^2.0.2
   money_formatter: ^0.0.3
@@ -47,12 +47,12 @@ dependencies:
   flutter_spinkit: ^5.1.0
   auto_size_text: ^3.0.0
   html: ^0.15.1
-  flutter_widget_from_html_core: ^0.9.0+2
+  flutter_widget_from_html_core: ^0.10.0
   flutter_rating_bar: ^4.0.1
   flutter_staggered_grid_view: ^0.6.2
   flutter_swiper_view: ^1.1.8
-  firebase_messaging: ^14.2.1
-  firebase_core: ^2.4.1
+  firebase_messaging: ^14.2.5
+  firebase_core: ^2.7.0
   flutter:
     sdk: flutter
   flutter_localizations:
