@@ -26,18 +26,16 @@ ThemeData darkTheme(ColorStyles darkColors) {
   }
 
   TextTheme darkTheme = getAppTextTheme(
-      appFont, defaultTextTheme.merge(_darkTextTheme(darkColors)));
+      appFont, defaultTextTheme.merge(_textTheme(darkColors)));
   return ThemeData(
     primaryColor: darkColors.primaryContent,
-    backgroundColor: darkColors.background,
-    colorScheme: ColorScheme.dark(),
     primaryColorDark: darkColors.primaryContent,
     brightness: Brightness.dark,
     focusColor: darkColors.primaryContent,
     scaffoldBackgroundColor: darkColors.background,
     appBarTheme: AppBarTheme(
         backgroundColor: darkColors.appBarBackground,
-        titleTextStyle: darkTheme.headline6!
+        titleTextStyle: darkTheme.titleLarge!
             .copyWith(color: darkColors.appBarPrimaryContent),
         iconTheme: IconThemeData(color: darkColors.appBarPrimaryContent),
         elevation: 1.0,
@@ -66,7 +64,7 @@ ThemeData darkTheme(ColorStyles darkColors) {
           TextStyle(color: darkColors.bottomTabBarLabelSelected),
       selectedItemColor: darkColors.bottomTabBarLabelSelected,
     ),
-    textTheme: darkTheme,
+    textTheme: darkTheme, colorScheme: ColorScheme.dark().copyWith(background: darkColors.background),
   );
 }
 
@@ -76,47 +74,13 @@ ThemeData darkTheme(ColorStyles darkColors) {
 |--------------------------------------------------------------------------
 */
 
-TextTheme _darkTextTheme(BaseColorStyles dark) {
-  final Color darkPrimaryContent = dark.primaryContent;
-  return TextTheme(
-    headline6: TextStyle(
-      color: darkPrimaryContent.withOpacity(0.8),
-    ),
-    headline5: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    headline4: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    headline3: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    headline2: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    headline1: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    subtitle2: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    subtitle1: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    overline: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    button: TextStyle(
-      color: darkPrimaryContent.withOpacity(0.8),
-    ),
-    bodyText2: TextStyle(
-      color: darkPrimaryContent.withOpacity(0.8),
-    ),
-    bodyText1: TextStyle(
-      color: darkPrimaryContent,
-    ),
-    caption: TextStyle(
-      color: darkPrimaryContent.withOpacity(0.8),
-    ),
+TextTheme _textTheme(ColorStyles colors) {
+  Color primaryContent = colors.primaryContent;
+  TextTheme textTheme = TextTheme().apply(displayColor: primaryContent);
+  return textTheme.copyWith(
+      titleLarge: TextStyle(color: primaryContent.withOpacity(0.8)),
+      labelLarge: TextStyle(color: primaryContent.withOpacity(0.8)),
+      bodySmall: TextStyle(color: primaryContent.withOpacity(0.8)),
+      bodyMedium: TextStyle(color: primaryContent.withOpacity(0.8))
   );
 }
