@@ -141,7 +141,7 @@ class _AccountDetailOrdersWidgetState extends State<AccountDetailOrdersWidget> {
                           formatStringCurrency(total: order.total),
                           style: Theme.of(context)
                               .textTheme
-                              .bodyText2!
+                              .bodyMedium!
                               .copyWith(fontWeight: FontWeight.w600),
                           textAlign: TextAlign.left,
                         ),
@@ -151,7 +151,7 @@ class _AccountDetailOrdersWidgetState extends State<AccountDetailOrdersWidget> {
                               trans("items"),
                           style: Theme.of(context)
                               .textTheme
-                              .bodyText1!
+                              .bodyLarge!
                               .copyWith(fontWeight: FontWeight.w600),
                           textAlign: TextAlign.left,
                         ),
@@ -168,7 +168,7 @@ class _AccountDetailOrdersWidgetState extends State<AccountDetailOrdersWidget> {
                             formatType: formatForDateTime(FormatType.time),
                           ),
                       textAlign: TextAlign.right,
-                      style: Theme.of(context).textTheme.bodyText1!.copyWith(
+                      style: Theme.of(context).textTheme.bodyLarge!.copyWith(
                             fontWeight: FontWeight.w400,
                           ),
                     ),
