@@ -19,7 +19,7 @@ class NoResultsForProductsWidget extends StatelessWidget {
         children: <Widget>[
           Text(
             trans("No results"),
-            style: Theme.of(context).textTheme.bodyText2,
+            style: Theme.of(context).textTheme.bodyMedium,
           ),
         ],
       );
