@@ -66,7 +66,7 @@ class _ProductDetailUpsellWidgetState extends State<ProductDetailUpsellWidget> {
               Text(
                 trans("${trans('You may also like')}…"),
                 style:
-                    Theme.of(context).textTheme.caption!.copyWith(fontSize: 18),
+                    Theme.of(context).textTheme.bodySmall!.copyWith(fontSize: 18),
                 textAlign: TextAlign.left,
               ),
             ],
