@@ -12,7 +12,7 @@ import 'package:nylo_framework/nylo_framework.dart';
 */
 
 // App Themes
-final List<BaseThemeConfig> appThemes = [
+final List<BaseThemeConfig<ColorStyles>> appThemes = [
   ThemeConfig.light(),
   ThemeConfig.dark(),
 ];
@@ -38,7 +38,7 @@ ColorStyles darkColors = DarkThemeColors();
 // Preset Themes
 class ThemeConfig {
   // LIGHT
-  static BaseThemeConfig light() => BaseThemeConfig(
+  static BaseThemeConfig<ColorStyles> light() => BaseThemeConfig<ColorStyles>(
         id: "default_light_theme",
         description: "Light theme",
         theme: lightTheme(lightColors),
@@ -46,7 +46,7 @@ class ThemeConfig {
       );
 
   // DARK
-  static BaseThemeConfig dark() => BaseThemeConfig(
+  static BaseThemeConfig<ColorStyles> dark() => BaseThemeConfig<ColorStyles>(
         id: "default_dark_theme",
         description: "Dark theme",
         theme: darkTheme(darkColors),
@@ -60,10 +60,10 @@ class ThemeConfig {
 
 // First add the colors which was created into the above section like the following:
 // Bright Colors
-  /// BaseColorStyles brightColors = BrightThemeColors();
+  /// ColorStyles brightColors = BrightThemeColors();
 
 // Next, uncomment the below:
-  /// static BaseThemeConfig bright() => BaseThemeConfig(
+  /// static BaseThemeConfig<ColorStyles> bright() => BaseThemeConfig<ColorStyles>(
   ///  id: "default_bright_theme",
   ///  description: "Bright theme",
   ///  theme: brightTheme(brightColors),
